+++ src/test/java/lotto/service/LottoServiceImplTest.java
@@ -6,14 +6,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import lotto.domain.entity.Lotto;
 import lotto.domain.entity.LottoResult;
 import lotto.domain.vo.Cash;
 import lotto.domain.vo.Rank;
-import lotto.repository.InMemoryLottoRepository;
-import lotto.repository.LottoRepository;
+import lotto.infrastructure.InMemoryLottoRepository;
+import lotto.domain.repository.LottoRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
